runs e2e tests only on push to master
