--- .github/workflows/nodejs.yml
@@ -1,43 +1,0 @@
-# This workflow will do a clean install of node dependencies, build the source code and run tests across different versions of node
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-nodejs-with-github-actions
-
-name: Node.js CI
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-
-    strategy:
-      matrix:
-        node-version: [12.x]
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Use Node.js ${{ matrix.node-version }}
-      uses: actions/setup-node@v1
-      with:
-        node-version: ${{ matrix.node-version }}
-    - run: npm ci
-    - run: npm run test:format
-    - run: npm run lint
-    - run: npm run build --if-present
-    - run: npm test
-      env:
-        CI: true
-        LRS_ENDPOINT: ${{ secrets.LRS_ENDPOINT_SCORMCLOUD }}
-        LRS_USERNAME: ${{ secrets.LRS_USERNAME_SCORMCLOUD }}
-        LRS_PASSWORD: ${{ secrets.LRS_PASSWORD_SCORMCLOUD }}
-    - run: npm test
-      env:
-        CI: true
-        LRS_ENDPOINT: ${{ secrets.LRS_ENDPOINT_VERACITY }}
-        LRS_USERNAME: ${{ secrets.LRS_USERNAME_VERACITY }}
-        LRS_PASSWORD: ${{ secrets.LRS_PASSWORD_VERACITY }}
-
